+++ main.py
@@ -1,5 +1,5 @@
-from decoding.decode import DecodePyzbar
-from localization.localize import LocalizeYolo
+from decoding.decode_pyzbar import DecodePyzbar
+from localization.localize_yolo import LocalizeYolo
 
 input_img = "dataset/PXL_20230826_081352782.jpg"
 det_path = "path to yolov7-stickers.onnx"
@@ -8,8 +8,4 @@ decoder = DecodePyzbar()
 find_barcodes = LocalizeYolo
 
 bounding_boxes = find_barcodes.yolov7_method(input_img, det_path)
-decoder.decode_p(input_img, bounding_boxes)
-
-
-# save_dir = "/Users/kseniia/Desktop/newtest"
-# crop_helper(save_dir, input_img, bounding_boxes)
+decoder.decode(input_img, bounding_boxes)
